Implement TwoTowerRankingEstimator.

+++ cpp/mindalpha/model_metric_buffer.cpp
@@ -19,44 +19,44 @@
 namespace mindalpha
 {
 
-void ModelMetricBuffer::UpdateBuffer(pybind11::array_t<int64_t> positive_buffer,
-                                     pybind11::array_t<int64_t> negative_buffer,
+void ModelMetricBuffer::UpdateBuffer(pybind11::array_t<double> positive_buffer,
+                                     pybind11::array_t<double> negative_buffer,
                                      pybind11::array_t<float> predictions,
-                                     pybind11::array_t<int64_t> labels)
+                                     pybind11::array_t<float> labels)
 {
     const size_t buffer_size = positive_buffer.size();
     const size_t instance_count = labels.size();
-    int64_t* const pos_buf = positive_buffer.mutable_data();
-    int64_t* const neg_buf = negative_buffer.mutable_data();
+    double* const pos_buf = positive_buffer.mutable_data();
+    double* const neg_buf = negative_buffer.mutable_data();
     const float* const preds = predictions.data();
-    const int64_t* const labs = labels.data();
+    const float* const labs = labels.data();
     for (size_t i = 0; i < instance_count; i++)
     {
         const float pred = preds[i];
-        const int64_t lab = labs[i];
+        const float lab = labs[i];
         const int64_t bucket = static_cast<int64_t>(pred * (buffer_size - 1));
-        if (lab > 0)
+        if (lab > 0.0)
             pos_buf[bucket] += lab;
-        else if (lab < 0)
+        else if (lab < 0.0)
             neg_buf[bucket] += -lab;
         else
-            neg_buf[bucket] += 1;
+            neg_buf[bucket] += 1.0;
     }
 }
 
-double ModelMetricBuffer::ComputeAUC(pybind11::array_t<int64_t> positive_buffer,
-                                     pybind11::array_t<int64_t> negative_buffer)
+double ModelMetricBuffer::ComputeAUC(pybind11::array_t<double> positive_buffer,
+                                     pybind11::array_t<double> negative_buffer)
 {
     const size_t buffer_size = positive_buffer.size();
-    const int64_t* const pos_buf = positive_buffer.mutable_data();
-    const int64_t* const neg_buf = negative_buffer.mutable_data();
+    const double* const pos_buf = positive_buffer.mutable_data();
+    const double* const neg_buf = negative_buffer.mutable_data();
     double auc = 0.0;
-    int64_t prev_pos_sum = 0;
-    int64_t pos_sum = 0;
-    int64_t neg_sum = 0;
+    double prev_pos_sum = 0;
+    double pos_sum = 0;
+    double neg_sum = 0;
     for (size_t i = 0; i < buffer_size; i++) {
-        const int64_t pos = pos_buf[i];
-        const int64_t neg = neg_buf[i];
+        const double pos = pos_buf[i];
+        const double neg = neg_buf[i];
         prev_pos_sum = pos_sum;
         pos_sum += pos;
         neg_sum += neg;
